Backend/requirements.txt
alembic==1.15.2
bcrypt==4.3.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.3
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.10
MarkupSafe==2.1.5
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.2
SQLAlchemy==2.0.40
urllib3==2.2.1
Werkzeug==3.0.3
